Add a PackageTest test for missing static files.
The test builds and installs a package dynamically only, then builds and
installs a package statically that depends on that dynamic package.

This demonstrates how cabal-install can fail when it ignores which build
artifacts are provided by a package, e.g. static and/or dynamic files.  Recent
changes add to the InstalledPackageInfo record so that installed packages are
more explicit in what is available, so that cabal-install is better able to
know when it should rebuild from source, when installed packages were not built
from config flags to build the required artifacts.