Track build artifacts in installed packages.
1) Extend the InstalledPackageInfo record with 5 records involving
configuration flags that affect which build artifacts (dynamic and static
files) are provided.  (This record corresponds to the ‘.conf’ files in
package-db directories, and Cabal-syntax provides an interface to this record.
Cabal-syntax is used as a dependency by ‘ghc-pkg’ and ‘cabal-install’, and old
Cabal-syntax implementations may produce an ‘Unknown field’ warning when used
with new ‘.conf’ files.)
2) Add a step to the validation phase of the modular resolver that also rules
out pre-installed package options that would introduce a missing artifact
condition, so that alternatives if available can be chosen instead, rather than
producing assignments for a build plan that would fail with ‘files missing’
(e.g. ‘….hi’ or ‘….dyn_hi’).