WIP: progress with an --enable-infer-unspecified flag.
I could finish this route, but perhaps instead I'll decide to simply have a
Bool for the various artifact kind flags that default to True, rather than my
more complicated original plan of having a Maybe Bool for these components,
where Nothing means it is unspecified, and the filesystem can be scanned.

Instead of it is unspecified, just assume it is ‘True’.  This is more in line
with the current behaviour.