WIP (TODO squash, track ?): phew, track debugs.
It works for my manual custom setup!  (git clone attoparsec and then build with
an Archlinux default dynamic-only system setup, manually adding IPI to
‘scientific’ to explicitly track that it doesn't have static arts.)

Now this needs an updated ghc-pkg, but for simplicity here I just manually
patch it in here.

Now this is mostly to track the debugs, although it has a couple or so minor
fixes I'll put back in.  This will be reverted, and probably soon I can start
wrapping things up.  (Woo!)