WIP: add to plan.json install-dirs for BuildAndInstall.
When elabBuildStyle is BuildInplaceOnly, the plan.json file includes
‘dist-dirs’, and this commit adds ‘install-dirs’ when elabBuildStyle is
BuildAndInstall, replacing the ‘TODO’ item.

TODO: In addition, the PackageTests ‘runPlanExe'’ helper now takes into account this
key-value pair when ‘dist-dirs’ is missing.  (This is needed for the
package-local flag fix to not cause the the T5782Diamond test to fail, since
the package-local flag fix causes the elabBuildStyle to change to
BuildAndInstall in the test, causing failure because of the technicality that
the helper assumes ‘dist-dirs’.  This change prevents this failure.)

(EDIT: TODO review this; plan.json has style: "local" but "bin-file: …"; is
this the right approach?)